Combine split_by_provider and dataset_substitution transforms into a more general from_datasets one
